cmd/compile: use staticuint64s for small constants
When converting integer constants to interfaces, we can instead use
pointers into staticuint64s for values that fit in a byte (bools,
byte/int8/uint8, and other integers in the range [0, 255]).

Do this by looking for stmps in convert.go, replacing them with
pointers into staticuint64s that are equivalent to their static data.
This leaves the stmps allocated, with the intent that they will be
eliminated during linking if unreferenced.

Counting, this optimization fires 928 times in std and cmd, most
commonly on error-implementing integers (like syscall.Errno) and parser
tokens.

This appears to be the last TODO in the issue, but there's more
improvement be had for the representation of static struct literals
containing small constants converted to interfaces.

Updates #37612

Change-Id: I17f8bfaa5447b0f4a3f68656f0057df08ec974d5