cmd/compile: use staticuint64s for small constants
When converting integer constants to interfaces, we can instead use
pointers into staticuint64s for values that fit in a byte (bools,
byte/int8/uint8, and other integers in the range [0, 255]).

This conversion is done early in order, since leaving literals in the
tree (as done in CL 223358) miscompiles struct literals, as convlit
treats literals as valid to be statically compiled, which for interface
conversions is not the case. In the aforementioned CL, this manifested
as a failure of issue31782.go.

Counting, this optimization fires 928 times in std and cmd, most
commonly on error-implementing integers (like syscall.Errno) and parser
tokens.

This appears to be the last TODO in the issue.

Fixes #37612

Change-Id: I17f8bfaa5447b0f4a3f68656f0057df08ec974d5