[dev.typeparams] cmd/compile: delay dictionary entry loading to later
Leave shape types in the IR until later in compilation, when we can
then resolve those types to the actual concrete types by looking up
in the dictionary.

Add dictionary scoping operations, to know which dictionary to use
(especially after inlining happens, when there might be multiple
dictionaries in a function).

Change-Id: I6ceda47667c7dd9ffffe936225419287f4ebdd9c