[dev.typeparams] cmd/compile/internal/types2: preserve untyped constants on the RHS of a shift expression
This is a port of CL 337529 with minor adjustments to two error calls and
without the extra test in check_test.go.

Instead document that the test is missing (enabling it will require some
adjustments to the testing framework in types2 which is slighly different
from the one in go/types).

Updates #47410
Updates #47243

Change-Id: I076e3f5d8cbcead3aeec7b3a8cef8453fd8133ed