cmd/compile: treat more address patterns as global
The previous CL added support for obvious fixed offsets of global
addresses, but this doesn't fix the current staticuint64s code, which
doesn't statically generate offsets during convert and instead relies on
lowering to produce the final address (after writebarrier).

Allow more address calculations involving global addresses to be treated
as global.

This isn't strictly required for the next CL, and could be omitted if
it isn't generally safe.

Updates #37612

Change-Id: I487622cc1fc5a4f81c78dfe9f7077a6c4201ecdc