[dev.fuzz] internal/fuzz: minimize interesting inputs
When a worker discovers an input that expands coverage, the worker
sends back to the coordinator. The coordinator will now send that
input back to the same worker for minimization. The coordinator will
only write the input to the corpus after minimization. When
minimizing, the worker will attempt to preserve at least one of the
coverage bits that made the input interesting.

TODO

- Coordinator status updates don't show that a worker is busy
  minimizing.
- Minimization can take a very long time for long inputs.
- Minimization cannot produce an empty string.
- Minimization should fail if no input (including the original input)
  preserves any of the new coverage bits. This can happen if a new
  coverage bit was set by a background goroutine.

Change-Id: Id26cb42c1e40b2ece87788781644fbd35e1c0339