cmd/internal/obj: change the R field of LSym to []*Reloc
Currently the R field of LSym is []Reloc, and generally we use the function
obj.Addrel to obtain the pointer of the R table entry to operate it. But since
the entry address of R is likely to change when appending new element to it,
then the previous pointer becomes invalid. This patch changes the R field to
[]*Reloc so that its pointer will not become invalid.

Change-Id: I8d679ce0dfa75dda35e09cfe97089fcd927bfae3