cmd/go/internal/base: add CmdFatalf, CmdErrorf, CmdLogf
CmdFatalf and CmdErrorf are like Fatalf and Errorf, but they print a
prefix like "go list: " or "go mod tidy: " based on the go subcommand
being run. This avoids the need to include such a prefix in calls to
Errorf and Fatalf.

Also added Logf and CmdLogf to replace fmt.Fprintf in a few places.

Change-Id: I790a38931e1ff65766d649ce2689320f9bf35a2d