cmd/go: use command prefix consistently in error messages
Previously, many error and log messages included a handwritten "go: "
prefix, or another prefix like "go list: " based on the current
subcommand. It was hard to get these to be presented consistently,
especially in common libraries like internal/modload.

This CL rewrites calls to fmt.Fprintf, base.Errorf, and base.Fatalf that
used a prefix to use base.CmdLogf, base.CmdErrorf, and base.CmdFatalf,
respectively. This was mostly done automatically with a refactoring
script, but hand edits were needed in a few places.

Change-Id: I7a9bfcbe6297ec93696130111974a53d74c7eb52