[dev.typeparams] cmd/compile, test: change -G default to 3 (types2 typechecker, generics)
Update the default value of Flag.G in flag.go. Must also update run.go
to understand that the value of the G flag is now 3.

-G=0 still works to use the old typechecker without generics.

Temporary change in plugin_test.go to try to debug failures. Seems
related to the build cache, but might be a -G=3 compiler bug that is
affecting the build code

Change-Id: Iabd099234b2bc8fafd708b2e99307eb4dc521966