[dev.typeparams] cmd/compile, test: change -G default to 3 (types2 typechecker, generics)
Update the default value of Flag.G in flag.go. Must also update run.go
to understand that the value of the G flag is now 3.

-G=0 still works to use the old typechecker without generics.

Fix smoketest.go to specify -G=1, since -G=3 doesn't handled some
operations on embedded fields in the test.

types2 can give us paths in Pos information that begin with $GOROOT.
Need to add a change in makeSrcPosBase() to look for the $GOROOT prefix,
since AbsFilename() doesn't deal correctly with input paths beginning
with $GOROOT. Otherwise, we can get wrong paths, that incorrectly depend
on the current directory (which was why the plugin test was failing).

Change-Id: Iabd099234b2bc8fafd708b2e99307eb4dc521966