cmd/internal/obj: correlate RestArgs of Prog with ordered position
At present, the RestArgs field of Prog is used to store a source or destination operand,
but there is no more detailed information to indicate its order. It's the third source
operand or destination operand on arm64 by default. The current implementation works
because we will not store more than one source or destination operand here. But the new
assembler sometimes needs to do this, for example:
	CSEL GT, R0, R19, R1
It has three source operands, p.From is GT, p.Reg is R0 and R19 is saved in p.RestArgs
with position obj.Source. In order to record which operand R0 corresponds to the machine
instruction, we need to convert p.Reg into an Addr and store it in p.RestArgs. If the
positions of this Addr and operand R19 are both obj.Source, then we can't distinguish them.

So this CL gives the operands in p.RestArgs more detailed position information, Source1,
Source2, Source3, Destination1 and Destination2 that can represent three source operands
and two destination operands. This CL also adds several functions to operate manipulate
the operand at the specified position.

For other architectures that do not require this information, Source1, Source2 and Source3
are the same, and Destination1 and Destination2 are the same.

Change-Id: Ic98a13bfdb5767c5d401e34c3af7d1d2592f22b0