[dev.typeparams] cmd/compile: change export version to 1.17 for testing
This is a temporary change. We will revert this back before the 1.18
release. We make this change now to simplify testing, since a lot of
tools will break on the new export version.

I will file a bug to revert this change and set the right export version
before we get close to release.

Change-Id: I0650fa753bb11229c71254d779dd61b5c1af9cdf