net/http: add ReuseConn hook
Add a ReuseConn hook to http.Transport and httptrace.ClientTrace.
The hook(s) are called when obtaining a new connection from an idle
pool, and may return true or false to indicate whether the connection
should be reused or discarded.

A ReuseConn hook may be used to implement a maximum lifetime for
connections, or a DNS-aware reuse policy which verifies that a
connection is still appropriate for a given DNS name.

Updates #23427.

Change-Id: I01552b5085aaca5ae970bdbc15de5298a885cf29