io: unexport internal methods
The methods on the pipe type don't need to be exported. Doing
so sets a bad precedent that it's OK to export methods to indicate
an internal public API, but it that's not a good idea in general because
exported methods increase cognitive load when reading code
because the reader needs to consider whether the exported method
might be used via some external interface or reflection.

Change-Id: Ib13f1b3f9fe0ff251628f31b776182a0953268ee