archive/tar: unexport internal methods
Many of the methods inside the archive/tar package don't need to be
exported. Doing so sets a bad precedent that it's OK to export methods
to indicate an internal public API, but it that's not a good idea in
general because exported methods increase cognitive load when reading
code because the reader needs to consider whether the exported method
might be used via some external interface or reflection.

This CL should have no externally visible behaviour changes at all.

Change-Id: I94a63de5e6a28e9ac8a283325217349ebce4f308