[dev.fuzz] testing: fix memory leak with sub-tests
Running a sub-test mutates the testing.T context since it inserts
the unique name of the test into testContext.match.subNames.
Since the Fuzz function is executed repeatedly, this causes
the subNames map to grow indefinitely with ever increasing
number of unique sub-names.

Fix the representation of matcher.subNames to only remember
unique input names, rather than all unique generated output names.
The memory requirements is only proportional to the number
of unique inputs, rather than the number of unique outputs.

Previously, the following:
	for i := 0; i < 1000; i++ {
		t.Run("name", ...)
	}
would have O(1000) memory, while this change makes it
such that it would have O(1) memory.

This CL makes the matcher logic more complex to detect when
a name looks auto-generated. The logic could be simplified
if we're willing to change the way auto-generated names
are derived. However, this CL aims to preserve the old behavior.

Fixes #44517

Change-Id: I564593cc066f6ccc09f9bc1b8e914920860c4b2b