[dev.fuzz] testing: fix memory leak with sub-tests
Running a sub-test mutates the testing.T context since it inserts
the unique name of the test into testContext.match.subNames.
Since the Fuzz function is executed repeatedly, this causes
the subNames map to grow indefinitely with ever increasing
number of unique sub-names.

Fix this issue by not calling t.context.match.fullName when fuzzing.
Create a bogus name by concatenating the parent name with the
sub-test name. It may not be unique, but that's okay,
it's never used to filter out sub-tests.

Fixes #44517

Change-Id: I564593cc066f6ccc09f9bc1b8e914920860c4b2b