cmd/go: draft of forward compatibility work
This CL exists simply to checkpoint and publish code I wrote
in September 2021 to play with the idea of extending Go's
forward compatibility by downloading newer Go distributions
as needed.

That idea has been discussed more recently as #55092,
as well as in my Compatibility talk (https://www.youtube.com/watch?v=v24wrd3RwGo).
This is the code I sketched a year ago to convince myself it could work.

This code is not for review. It is only to make a copy public for reference.

This CL is also incomplete - I appear to have lost the code that
took care of downloading a Go distribution and turning it into
a module zip file, which was cmd/go/internal/godist.

DO NOT REVIEW

Change-Id: If54507e57709e135b3e5e4a335ad4fff83e01247