cmd/compile: fix CONVIFACE case converting interface to empty interface
We need an extra case in convertToDictionary. In the case of an operand
which is an interface, we don't want to get the run-time type from the
dictionary (which would be the run-time type of the interface). We want
to get the run-time type from the run-time interface value, or
equivalently, we want to do a type-assert to the empty interface.

Change-Id: I414247210168153151272fab198bfe82ad7b1567