cmd/go/internal/modload: move BuildMod into State
Previously, cfg.BuildMod was used both for the value of the -mod flag
and for the inferred mode when it wasn't set. We also had
BuildModExplicit to track whether it was set explicitly and
BuildModReason to explain the inferred mode (both affecting error
messages). These globals were used extensively in modload and
modfetch, and there were several special cases that changed behavior
based on command name and whether AllowMissingModuleImports was
called.

This change adds modload.State.Mod, which is now the canonical value
to check. cfg.BuildMod now ONLY tracks the value of the -mod flag, and
BuildModExplicit is deleted. BuildModReason is moved to
State.ModReason. This change also removes AllowMissingModuleImports;
commands should only rely on State.Mod.

This required extensive plumbing to get State everywhere State.Mod was
needed. Hopefully, the remaining changes will be simpler.

For #40775

Change-Id: I4b66eee1267bd1f915ecdbee17a7cff314ceab11