cmd/compile: change types2 typename format to be like types1
Change types2.TypeString() to output 'interface {' and 'struct {',
rather than current 'interface{' and 'struct{', so that it is compatible
with the way types1 prints types (via types.Type.String()). Main changes
in typestring.go.

This allows the removal of code in makeInstName1() and InstTypeName()
(related to 'inteface {' and 'interface{' difference) for -G=3 mode, and
fixes issue 47877, which results from the 'struct{' and 'struct {'
difference.

Fixes #47877.

Change-Id: I5fa563d0867a8468dddd4cb5fedc2844c1505efc