cmd/compile: fix parameterized empty interfaces
type I[T any] interface{}

This is an empty interface, but it has a type parameter.
We need to distinguish that from an empty interface that is not parameterized.

That means when doing type substitution on an empty interface with
parameters, we need to make a new one.

Similar case to tstruct, above.

Change-Id: I23ad9f21d2c4ef675bf3f7d84899d9e4919d05e7