cmd/compile: fix reference to generic type needed by crawler
This problem happens when you create a new local type that includes an
imported generic type. In that case, the imported generic type is does
not have a declaration in the local package, so it is not necessarily
created in types1.

To fix this issue, we just need to add a call to g.obj() for the base
generic type, to make sure that it will exist if needed later in the
compilation or for the crawler during export.

Fixes #47514

Change-Id: Ie756578f07ad0007de8a88ae909cf7534a22936e